Tutorial 17: Texturing (Working)

+++ directx/Vector2.h
@@ -4,10 +4,10 @@
 
 class Vector2 {
 	public:
-		int x = 0, y = 0; 
+		float x = 0, y = 0;
 
 		Vector2() : x(0), y(0) {}
-		Vector2(int x, int y) : x(x), y(y) {}
+		Vector2(float x, float y) : x(x), y(y) {}
 		Vector2(const Vector2& point) : x(point.x), y(point.y) {}
 		~Vector2() {}
